requirements.txt
mysql-connector-python==9.1.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
python-dateutil==2.9.0.post0
pytz==2024.2
shiboken6==6.8.0.2
six==1.16.0
tzdata==2024.2
